requirements.txt
alembic==1.17.2
bcrypt==5.0.0
bleach==6.3.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-CKEditor==1.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mysqlclient==2.2.7
packaging==25.0
pycparser==2.23
PyMySQL==1.1.2
python-dotenv==1.2.1
SQLAlchemy==2.0.45
typing_extensions==4.15.0
utils==1.0.2
webencodings==0.5.1
Werkzeug==3.1.4
WTForms==3.2.1
